* Fix - expected change to fix jacoco report issue and better package name and location. Second part: refactor Enumeration for SuiteResultType.
