+++ testApi/src/test/scala/africa/absa/testing/scapi/reporter/StdOutReporterTest.scala
@@ -17,7 +17,8 @@
 package africa.absa.testing.scapi.reporter
 
 import africa.absa.testing.scapi.AssertionException
-import africa.absa.testing.scapi.model.suite.{SuiteResult, SuiteResultType}
+import africa.absa.testing.scapi.model.suite.SuiteResult
+import africa.absa.testing.scapi.model.suite.types.SuiteResultType
 import munit.FunSuite
 
 import java.io.ByteArrayOutputStream
@@ -26,20 +27,20 @@ import scala.util.{Failure, Success}
 class StdOutReporterTest extends FunSuite {
 
   val successTestResults: List[SuiteResult] = List(
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 1",
       name = "Test 1",
       result = Success(()),
       duration = Some(100L),
       categories = Some("Category 1")),
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 1",
       name = "Test 2",
       result = Success(()),
       duration = Some(200L),
       categories = Some("Category 2")
     ),
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 2",
       name = "Test 1",
       result = Success(()),
@@ -48,19 +49,19 @@ class StdOutReporterTest extends FunSuite {
   )
 
   val mixedSuccessTestResults: List[SuiteResult] = List(
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 1",
       name = "Test 1",
       result = Success(()),
       duration = Some(100L),
       categories = Some("Category 1")),
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 1",
       name = "Test 2",
       result = Failure(AssertionException("Error message")),
       duration = Some(200L),
       categories = Some("Category 2")),
-    SuiteResult(SuiteResultType.TEST_SUITE,
+    SuiteResult(SuiteResultType.TEST_SET,
       suiteName = "Suite 2",
       name = "Test 1",
       result = Success(()),
